Add comprehensive tests for CPU backend correctness and distribution autograd
- Introduced a new test suite for CPU backend correctness in `test_release_audit_cpu.cpp`, covering various functionalities including `isin`, `index_select`, vector norms, and layer normalization.
- Added tests to ensure proper gradient flow in distribution log probabilities and reparameterization for Gumbel, Gamma, Beta, Chi2, StudentT, Dirichlet, Weibull, and HalfNormal distributions in `test_distributions_autograd.py`.
- Implemented tests for DLPack version negotiation in `test_dlpack_versioned.py`, ensuring compatibility with NumPy.
- Verified functional embedding correctness in `test_functional_embedding.py`, ensuring gradients flow correctly to supplied weights.
- Enhanced zero-copy functionality in `test_numpy_zero_copy.py`, confirming shared memory behavior for contiguous and non-contiguous NumPy arrays.

+++ src/ops/math.cpp
@@ -946,7 +946,10 @@ auto logit(const Tensor& input, double eps) -> Tensor {
     // With eps > 0: clamp x to [eps, 1-eps] first
     Tensor x = input;
     if (eps > 0) {
-        x = clamp(x, static_cast<float>(eps), static_cast<float>(1.0 - eps));
+        // clamp takes double bounds; casting to float here corrupted the
+        // [eps, 1-eps] window for Float64 inputs (1-1e-6 is not representable
+        // in float, shifting the result near the boundary by ~1e-2).
+        x = clamp(x, eps, 1.0 - eps);
     }
     Tensor one = ones_like(x);
     return log(x / (one - x));
